Move common.dbos to client.dtos
