+++ ridi_oauth2/client/grant/base.py
@@ -6,7 +6,7 @@ import requests
 from ridi_oauth2.client.dtos import AuthorizationServerInfo, ClientInfo
 from ridi_oauth2.client.exceptions import AuthorizationException, OAuthFailureException
 from ridi_oauth2.common.constants import HttpStatusCode
-from ridi_oauth2.common.dtos import TokenData
+from ridi_oauth2.client.dtos import TokenData
 
 
 class BaseGrant:
