+++ tests/tests_ridi_django_oauth2_resource/test_authentication_middleware.py
@@ -1,18 +1,14 @@
-import json
 import time
 from unittest.mock import Mock
 
 import jwt
-import requests_mock
 from django.contrib.auth import get_user_model
 from django.contrib.auth.models import AnonymousUser
-from django.http import HttpResponse
 from django.test import TestCase
 
 from ridi_django_oauth2_resource.config import RidiOAuth2Config
 from ridi_django_oauth2_resource.middlewares import AuthenticationMiddleware
 from ridi_django_oauth2_resource.response import HttpUnauthorizedResponse
-from ridi_oauth2.common.dtos import TokenData
 
 
 class AuthenticationMiddlewareTestCase(TestCase):
