+++ tests/tests_ridi_oauth2/test_token_dtos.py
@@ -1,9 +1,8 @@
 import unittest
 from datetime import datetime
 
-from ridi_oauth2.common.dtos import TokenData
+from ridi_oauth2.client.dtos import TokenData
 from ridi_oauth2.introspector.dtos import AccessTokenInfo
-
 
 class TokenTestCase(unittest.TestCase):
     def setUp(self):
